fixing 10

+++ pointers_arrays_strings/100-atoi.c
@@ -33,9 +33,9 @@ int _atoi(char *s)
 		{
 			started = 1;
 			num = num * 10 + (*s - '0');
-			if (num > 2147483647 && sign == 1)
+			if (num > (unsigned int)(2147483647U) && sign == 1)
 				return (-1);
-			if (num > 2147483648U && sign == -1)
+			if (num > (unsigned int)(2147483648U) && sign == -1)
 				return (0);
 		}
 		else if (started)
